NPM-PACKAGE: Move @microsoft/kiota to ESM-only
Lifts PR #7441 (deinok) onto current main, with fixes so the build,
unit tests, lint and extension webpack bundling all pass.

- Ship ESM only (drop dual CJS/ESM build, tsconfig.cjs/esm.json)
- Relocate remaining root modules into lib/ so the imported runtime.json
  is emitted to dist by tsc (removes the copy-files.js script)
- Use explicit .js import specifiers under module/moduleResolution nodenext
- Point CI runtimeFilePath at vscode/packages/npm-package/lib/runtime.json

Fixes over the original PR:
- Add .js extension to vscode-jsonrpc/node imports so they resolve under
  nodenext ESM (was TS2307; also unblocked cascading implicit-any errors)
- Update tests/unit/install.spec.ts import to ../../lib/install
- Preserve current main package.json version and dependencies
