+++ vscode/packages/npm-package/lib/generatePlugin.ts
@@ -1,9 +1,9 @@
-import * as rpc from "vscode-jsonrpc/node";
+import * as rpc from "vscode-jsonrpc/node.js";
 
-import connectToKiota from "../connect";
-import { KiotaPluginType, GeneratePluginResult, ConsumerOperation, GenerationConfiguration, KiotaLogEntry, PluginAuthType } from "../types";
+import connectToKiota from "./connect.js";
+import { KiotaPluginType, GeneratePluginResult, ConsumerOperation, GenerationConfiguration, KiotaLogEntry, PluginAuthType } from "./types.js";
 import * as path from "path";
-import { checkForSuccess } from "../utils";
+import { checkForSuccess } from "./utils.js";
 
 export interface PluginGenerationOptions {
   descriptionPath: string;
