+++ vscode/packages/npm-package/lib/migrateFromLockFile.ts
@@ -1,7 +1,7 @@
-import * as rpc from "vscode-jsonrpc/node";
+import * as rpc from "vscode-jsonrpc/node.js";
 
-import connectToKiota from "../connect";
-import { KiotaLogEntry } from "../types";
+import connectToKiota from "./connect.js";
+import { KiotaLogEntry } from "./types.js";
 
 /**
  * Migrates data from a lock file located in the specified directory.
