+++ vscode/packages/npm-package/lib/updateClients.ts
@@ -1,7 +1,7 @@
-import * as rpc from "vscode-jsonrpc/node";
+import * as rpc from "vscode-jsonrpc/node.js";
 
-import connectToKiota from "../connect";
-import { KiotaLogEntry } from "../types";
+import connectToKiota from "./connect.js";
+import { KiotaLogEntry } from "./types.js";
 
 export interface UpdateClientsConfiguration {
   cleanOutput: boolean;
