+++ vscode/packages/npm-package/tests/unit/generatePlugin.spec.ts
@@ -1,6 +1,6 @@
 import { generatePlugin } from '../../lib/generatePlugin';
-import { KiotaLogEntry } from '../../types';
-import { KiotaPluginType } from '../../types';
+import { KiotaLogEntry } from '../../lib/types';
+import { KiotaPluginType } from '../../lib/types';
 import { setupKiotaStubs } from './stubs.util';
 
 describe("generate plugin", () => {
