--- vscode/packages/npm-package/tsconfig.cjs.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist/cjs",
-    "module": "CommonJS",
-    "declaration": true,
-    "declarationDir": "./dist/cjs/types"
-  }
-}
