--- vscode/packages/npm-package/tsconfig.esm.json
@@ -1,9 +1,0 @@
-{
-  "extends": "./tsconfig.json",
-  "compilerOptions": {
-    "outDir": "./dist/esm",
-    "module": "ESNext",
-    "declaration": true,
-    "declarationDir": "./dist/esm/types"
-  }
-}
