+++ vscode/packages/npm-package/tsconfig.json
@@ -1,16 +1,19 @@
 {
+  "$schema": "https://www.schemastore.org/tsconfig.json",
   "compilerOptions": {
+    "declaration": true,
+    "declarationMap": true,
     "forceConsistentCasingInFileNames": true,
-    "module": "ESNext",
-    "rootDir": ".",
-    "outDir": "./dist",
+    "module": "nodenext",
+    "moduleResolution": "nodenext",
+    "rootDir": "./lib/",
+    "outDir": "./dist/",
     "strict": true,
     "types": [
       "node",
       "jest"
     ],
-    "moduleResolution": "node",
-    "target": "ESNext",
+    "target": "esnext",
     "lib": [
       "ESNext"
     ],
@@ -20,9 +23,7 @@
     "allowSyntheticDefaultImports": true,
     "skipLibCheck": true,
   },
-  "exclude": [
-    "dist",
-    "node_modules",
-    "tests"
+  "include": [
+    "./lib/**/*.ts"
   ]
 }
